requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
pypdf==3.17.4
python-pptx==0.6.23
python-docx==1.1.0
pytesseract==0.3.10
Pillow==10.1.0
unstructured==0.11.6
openai==1.3.7
ollama==0.1.4
pydantic==2.5.0

